+++ tests/test_keeper.py
@@ -1,19 +1,12 @@
-from platform import system
-
-import pytest
-
-
-@pytest.mark.skipif(
-    system() == "Windows", reason="Metaflow is not supported on Windows"
+from resource_tracker.keeper import (
+    get_instance_price,
+    get_recommended_cloud_servers,
+    keeper_request,
 )
+
+
 def test_keeper_request():
     """Test Keeper API requests."""
-    from metaflow_extensions.resource_tracker.plugins.cards.tracked_resources.helpers import (
-        get_instance_price,
-        get_recommended_cloud_servers,
-        keeper_request,
-    )
-
     assert keeper_request("/healthcheck") is not None
     assert get_instance_price("aws", "us-east-1", "t3.micro") is not None
     assert get_recommended_cloud_servers(1, 1024, n=1) is not None